Revert "Weekly reporting"
