+++ invoice_checking/urls.py
@@ -17,12 +17,9 @@ urlpatterns = [
     url(r'^project/(?P<project_id>[0-9A-Fa-f-]+)$', invoices.views.project_details, name='project'),
     url(r'^project/(?P<project_id>[0-9A-Fa-f-]+)/charts$', invoices.views.project_charts, name='project_charts'),
     url(r'^invoice/(?P<invoice_id>[0-9A-Fa-f-]+)/hours$', invoices.views.invoice_hours, name='invoice_hours'),
-    url(r'^weekly_report/(?P<weekly_report_id>[0-9A-Fa-f-]+)/hours$', invoices.views.weekly_report_hours, name='weekly_report_hours'),
     url(r'^invoice/(?P<invoice_id>[0-9A-Fa-f-]+)/charts$', invoices.views.invoice_charts, name='invoice_charts'),
-    url(r'^invoice/(?P<invoice_id>[0-9A-Fa-f-]+)/pdf$', invoices.views.get_pdf, name="get_invoice_pdf"),
+    url(r'^invoice/(?P<invoice_id>[0-9A-Fa-f-]+)/pdf/(?P<pdf_type>.+)$', invoices.views.get_pdf, name="get_pdf"),
     url(r'^invoice/(?P<invoice_id>[0-9A-Fa-f-]+)$', invoices.views.invoice_page, name="invoice"),
-    url(r'^weekly_report/(?P<weekly_report_id>[0-9A-Fa-f-]+)$', invoices.views.weekly_report_page, name="weekly_report"),
-    url(r'^weekly_report/(?P<weekly_report_id>[0-9A-Fa-f-]+)/pdf$', invoices.views.get_pdf, name="get_weekly_pdf"),
     url(r'^amazon_invoice/(?P<linked_account_id>[0-9A-Fa-f-]+)/(?P<year>[0-9]{4})/(?P<month>[0-9]{1,2})$', invoices.views.amazon_invoice, name="amazon_invoice"),
     url(r'^amazon$', invoices.views.amazon_overview, name="amazon_overview"),
     url(r'^hours$', invoices.views.hours_list, name="hours_list"),
@@ -31,7 +28,6 @@ urlpatterns = [
     url(r'^people/hourmarkings$', invoices.views.people_hourmarkings, name='people_hourmarkings'),
     url(r'^people/charts$', invoices.views.people_charts, name='people_charts'),
     url(r'^person/(?P<user_guid>[0-9A-Fa-f-]+)/(?P<year>[0-9]{4})/(?P<month>[0-9]{1,2})', invoices.views.person_details_month, name='person'),
-    url(r'^personweek/(?P<user_guid>[0-9A-Fa-f-]+)/(?P<year>[0-9]{4})/(?P<week>[0-9]{1,2})',invoices.views.person_details_week, name='personweek'),
     url(r'^person/(?P<user_guid>[0-9A-Fa-f-]+)', invoices.views.person_details, name='person_details'),
     url(r'^admin/', include(admin.site.urls)),
     url(r'^queue_update$', invoices.views.queue_update, name="queue_update"),
