+++ invoices/templates/invoice_hours.html
@@ -13,7 +13,35 @@
 
     <h2>{{ invoice.full_name }}: {{ invoice.date }}</h2>
 
-    {%  include "snippets/report_hours.html" %}
+    <table class="table table-striped table-condensed">
+      <thead>
+        <tr>
+          <th>Date</th>
+          <th>Person</th>
+          <th>Phase</th>
+          <th>Category</th>
+          <th>Billable</th>
+          <th>Notes</th>
+          <th class="number-column">Hours</th>
+          <th class="number-column">Rate</th>
+          <th class="number-column">Price</th>
+      </thead>
+      <tbody>
+        {% for entry in entries %}
+        <tr>
+          <td class="nowrap-column">{{ entry.date }}</td>
+          <td class="nowrap-column">{% if entry.user_m %}<a href="{% url "person" entry.user_m.guid entry.date.year entry.date.month %}">{{ entry.user_name }}</a>{% else %}{{ entry.user_name }}{% endif %}</td>
+          <td class="nowrap-column">{{ entry.phase_name }}</td>
+          <td class="nowrap-column">{{ entry.category }}</td>
+          <td>{{ entry.calculated_is_billable }}</td>
+          <td>{{ entry.notes }}</td>
+          <td class="number-column">{{ entry.incurred_hours|floatformat:2 }}h</td>
+          <td class="number-column">{{ entry.bill_rate|floatformat:2 }}&euro;/h</td>
+          <td class="number-column">{{ entry.incurred_money|floatformat:2 }}&euro;</td>
+        </tr>
+        {% endfor %}
+      </tbody>
+    </table>
   </div>
 </div>
 {% endblock %}
