+++ invoices/templates/invoice_page.html
@@ -31,11 +31,8 @@
           </ul>
         </div>
         {% endif %}
-        {%  if recent_weekly_report %}
-          <a class="btn btn-info" href="{% url 'weekly_report' recent_weekly_report.weekly_report_id %}" role="button">Weekly report</a>
-        {%  endif %}
-          <a class="btn btn-info" href="{% url 'invoice_charts' invoice.invoice_id %}" role="button">Charts</a>
-        <a class="btn btn-info" href="{% url 'get_invoice_pdf' invoice.invoice_id %}" role="button">Download hours</a>
+        <a class="btn btn-info" href="{% url 'invoice_charts' invoice.invoice_id %}" role="button">Charts</a>
+        <a class="btn btn-info" href="{% url 'get_pdf' invoice.invoice_id 'hours' %}" role="button">Download hours</a>
         <a class="btn btn-info" href="{% url 'invoice_hours' invoice.invoice_id %}" role="button">View hours</a>
         {% if invoice.project_m %}
         <a class="btn btn-info" href="https://app.10000ft.com/viewproject?id={{ invoice.project_m.project_id }}" role="button">View project in 10000ft</a>
@@ -56,7 +53,7 @@
       <div class="col-md-8">
         {% include "snippets/invoice.html" %}
         {% if last_month_invoice %}
-          {% include "snippets/changes_between_reports.html" with  month_or_week="month" last_report=last_month_invoice this_report=invoice diff_last_report=diff_last_month %}
+          {% include "snippets/last_month_invoice.html" %}
         {% endif %}
         {% include "snippets/incorrect_data.html" %}
       </div>
