+++ invoices/templates/project_details.html
@@ -29,13 +29,8 @@
 
 
 <div class="row">
-  <div class="col-md-8">
+  <div class="col-md-12">
     {% render_table invoices %}
-  </div>
-  <div class="col-md-4">
-    <div class="well">
-      {% include "snippets/weekly_report_enable_notifications.html" %}
-    </div>
   </div>
 </div>
 {% endblock %}
