--- invoices/templates/snippets/changes_between_reports.html
@@ -1,37 +1,0 @@
-<h2>Changes since last {{ month_or_week }}</h2>
-<table class="table">
-  <thead>
-    <tr>
-      <th></th>
-      <th class="number-column">Hours</th>
-      <th class="number-column">Bill rate</th>
-      <th class="number-column">Money</th>
-    </tr>
-  </thead>
-  <tbody>
-    <tr>
-      <td>Last {{ month_or_week }}</td>
-      <td class="number-column">{{ last_report.incurred_hours|floatformat:2 }}h</td>
-      <td class="number-column">{{ last_report.bill_rate_avg|floatformat:2 }}&euro;</td>
-      <td class="number-column">{{ last_report.incurred_money|floatformat:2 }}&euro;</td>
-    </tr>
-    <tr>
-      <td>This {{ month_or_week }}</td>
-      <td class="number-column">{{ this_report.incurred_hours|floatformat:2 }}h</td>
-      <td class="number-column">{{ this_report.bill_rate_avg|floatformat:2 }}&euro;</td>
-      <td class="number-column">{{ this_report.incurred_money|floatformat:2 }}&euro;</td>
-    </tr>
-    <tr>
-      <td>Difference</td>
-      <td class="number-column">{{ diff_last_report.hours.diff|floatformat:2 }}h</td>
-      <td class="number-column">{{ diff_last_report.bill_rate_avg.diff|floatformat:2 }}&euro;</td>
-      <td class="number-column">{{ diff_last_report.money.diff|floatformat:2 }}&euro;</td>
-    </tr>
-    <tr>
-      <td>Diff %</td>
-      <td class="number-column">{{ diff_last_report.hours.percentage|floatformat:2 }}%</td>
-      <td class="number-column">{{ diff_last_report.bill_rate_avg.percentage|floatformat:2 }}%</td>
-      <td class="number-column">{{ diff_last_report.money.percentage|floatformat:2 }}%</td>
-    </tr>
-  </tbody>
-</table>
