+++ invoices/templates/snippets/invoice.html
@@ -39,7 +39,6 @@
                 </td>
             </tr>
 
-        {% include "snippets/project_hours.html" %}
 
 
             {% for phase_name, phase in phases.items %}
