--- invoices/templates/snippets/person_hour_dropdown_month.html
@@ -1,3 +1,0 @@
-{% for entries_month in months %}
-    <li><a href="{% url "person" person.guid entries_month.year entries_month.month %}">{{ entries_month }}</a></li>
-{%  endfor %}
