--- invoices/templates/snippets/project_hours.html
@@ -1,89 +1,0 @@
-{% load humanize %}
-
-{% for phase_name, phase in phases.items %}
-    <tr class="heading">
-        <td>
-            {{ phase_name }}
-        </td>
-
-        <td>
-          Bill rate
-        </td>
-        <td>
-            Hours
-        </td>
-        <td>
-            Price
-        </td>
-    </tr>
-      {% if phase.entries %}
-        {% for entry_name, price in phase.entries.items %}
-        <tr class="item">
-            <td>{% if price.aws_id and invoice %}<a href="{% url "amazon_invoice" price.aws_id invoice.year invoice.month %}">{{ entry_name }}</a>
-              {% else %}{{ entry_name }}{% endif %}</td>
-            <td>-</td>
-            <td>-</td>
-            <td>
-              {% if not phase.billable %}<s>{% endif %}
-                {% if price.decimals %}
-                  {{ price.price|floatformat:price.decimals|intcomma }} {{ price.currency}}
-                {% else %}
-                  {{ price.price|floatformat:2|intcomma }} {{ price.currency}}
-                {% endif %}
-              {% if not phase.billable %}</s>{% endif %}
-            </td>
-        </tr>
-        {% endfor %}
-      {% else %}
-      {% for person_email, person_details in phase.users.items %}
-        {% for bill_rate, details in person_details.entries.items %}
-          {% if details.incurred_hours > 0 %}
-            <tr class="item">
-                <td>
-                    {% if person_details.user_m and not customer_preview %}
-                        {%  if invoice %}
-                            <a href="{% url 'person' person_details.user_m.guid invoice.year invoice.month %}">{{ person_details.user_name }}</a>
-                        {%  else %}
-                            <a href="{% url 'personweek' person_details.user_m.guid weekly_report.year weekly_report.week %}">{{ person_details.user_name }}</a>
-                        {%  endif %}
-                    {% else %}{{ person_details.user_name }}
-                    {% endif %}
-                </td>
-
-                <td>
-                  {{ bill_rate|floatformat:2 }} EUR
-                </td>
-                <td>
-                  {{ details.incurred_hours|floatformat:2 }}h
-                </td>
-                <td>
-                  {% if not phase.billable %}<s>{% endif %}
-                  {{ details.incurred_money|floatformat:2|intcomma }} EUR
-                  {% if not phase.billable %}</s>{% endif %}
-                </td>
-            </tr>
-          {% endif %}
-        {% endfor %}
-      {% endfor %}
-      <tr class="total">
-        <td></td>
-        <td></td>
-        <td>{{ phase.incurred_hours }}h</td>
-        <td>{{ phase.incurred_money|floatformat:2|intcomma }} EUR</td>
-      </tr>
-      {% endif %}
-    {% endfor %}
-
-    {% for total_row in total_rows.values %}
-    <tr class="total">
-        <td>{{ total_row.description }}</td>
-        <td>{% if total_row.bill_rate_avg %}{{ total_row.bill_rate_avg|floatformat:2 }} {{ total_row.currency }}{% endif %}</td>
-        <td>
-          {% if total_row.incurred_hours %}{{ total_row.incurred_hours|floatformat:2 }}h{% endif %}
-        </td>
-
-        <td>
-          {{ total_row.incurred_money|floatformat:2|intcomma }} {{ total_row.currency }}
-        </td>
-    </tr>
-{% endfor %}
