--- invoices/templates/snippets/weekly_report.html
@@ -1,44 +1,0 @@
-{% load humanize %}
-
-    <div class="invoice-box">
-        <table cellpadding="0" cellspacing="0">
-            <tr class="top">
-                <td colspan="4">
-                    <table>
-                        <tr>
-                            <td class="title">
-                                <img src="https://solinor.fi/wp-content/themes/twentyfifteen-child/images/logo-solinor@2x.png" style="width:100%; max-width:300px;">
-                            </td>
-
-                            <td>
-                                Weekly report #: {{ weekly_report.week }}<br>
-                                Created: {{ today }}<br>
-                            </td>
-                        </tr>
-                    </table>
-                </td>
-            </tr>
-
-            <tr class="information">
-                <td colspan="4">
-                    <table>
-                        <tr>
-                            <td>
-                                Solinor Oy<br>
-                                Teollisuuskatu 21<br>
-                                00510 Helsinki
-                            </td>
-
-                            <td>
-                                {{ weekly_report.project_m.client }}<br>
-                                {{ weekly_report.project_m.name }}
-                            </td>
-                        </tr>
-                    </table>
-                </td>
-            </tr>
-
-        {% include "snippets/project_hours.html" with report=weekly_report url="personweek" %}
-
-        </table>
-    </div>
