--- invoices/templates/snippets/weekly_report_accept.html
@@ -1,10 +1,0 @@
-<form method="post">
-  {% csrf_token %}
-  <h3>Team lead</h3>
-  <input type="checkbox" id="weeklyReportChecked" name="weeklyReportChecked" {% if weekly_report.is_approved %}checked="checked"{% endif %}/>
-  <label for="weeklyReportChecked">I have verified that the weekly report is OK</label>
-  <button type="submit" class="btn btn-primary">Save</button>
-  {% if form_data %}
-    <p>Last submitted by {{ form_data.user }} at {{ form_data.timestamp }}</p>
-  {% endif %}
-</form>
