--- invoices/templates/snippets/weekly_report_enable_notifications.html
@@ -1,10 +1,0 @@
-<form method="post">
-  {% csrf_token %}
-  <h3>Team lead</h3>
-  <input type="checkbox" id="weeklyReportEnableNotifications" name="weeklyReportEnableNotifications" {% if project.enable_weekly_notifications %}checked="checked"{% endif %}/>
-  <label for="weeklyReportEnableNotifications">Enable weekly report slack notifications</label>
-  <button type="submit" class="btn btn-primary">Save</button>
-  {% if form_data %}
-    <p>Last submitted by {{ form_data.user }} at {{ form_data.timestamp }}</p>
-  {% endif %}
-</form>
