+++ invoices/views.py
@@ -3,7 +3,6 @@
 import datetime
 import json
 import copy
-import calendar
 from collections import defaultdict
 
 import redis
@@ -21,12 +20,11 @@ from django.db.models.functions import TruncMonth
 
 from django_tables2 import RequestConfig
 
-from invoices.models import HourEntry, Invoice, WeeklyReport, Comments, DataUpdate, FeetUser, Project, AuthToken, \
-    InvoiceFixedEntry, ProjectFixedEntry, AmazonInvoiceRow, AmazonLinkedAccount, WeeklyReportComments
+from invoices.models import HourEntry, Invoice, Comments, DataUpdate, FeetUser, Project, AuthToken, InvoiceFixedEntry, ProjectFixedEntry, AmazonInvoiceRow, AmazonLinkedAccount
 from invoices.filters import InvoiceFilter, ProjectsFilter, CustomerHoursFilter, HourListFilter
-from invoices.pdf_utils import generate_hours_pdf_for_invoice, generate_hours_pdf_for_weekly_report
+from invoices.pdf_utils import generate_hours_pdf_for_invoice
 from invoices.tables import HourListTable, CustomerHoursTable, FrontpageInvoices, ProjectsTable, ProjectDetailsTable
-from invoices.invoice_utils import generate_amazon_invoice_data, calculate_entry_stats, calculate_weekly_entry_stats, get_aws_entries, compare_invoices
+from invoices.invoice_utils import generate_amazon_invoice_data, calculate_entry_stats, get_aws_entries
 import invoices.date_utils as date_utils
 from invoices.chart_utils import gen_treemap_data_projects, gen_treemap_data_users
 
@@ -39,7 +37,6 @@ def validate_auth_token(auth_token):
         if token.valid_until > timezone.now():
             raise Http404()
     return token
-
 
 @login_required
 def amazon_overview(request):
@@ -78,7 +75,6 @@ def amazon_overview(request):
         "linking_data_json": json.dumps(linking_data),
     }
     return render(request, "amazon_overview.html", context)
-
 
 @login_required
 def amazon_invoice(request, linked_account_id, year, month):
@@ -127,7 +123,6 @@ def customer_view(request, auth_token):
         "auth_token": auth_token,
     }
     return render(request, "customer_main.html", context)
-
 
 def customer_view_invoice(request, auth_token, year, month):
     token = validate_auth_token(auth_token)
@@ -210,7 +205,6 @@ def person_details(request, user_guid):
 
     return render(request, "person_details.html", {"entries": entries, "person": person, "calendar_charts": calendar_charts, "months": months, "treemap_charts": treemaps})
 
-
 @login_required
 def person_details_month(request, year, month, user_guid):
     year = int(year)
@@ -218,29 +212,7 @@ def person_details_month(request, year, month, user_guid):
     person = get_object_or_404(FeetUser, guid=user_guid)
     entries = person.hourentry_set.exclude(incurred_hours=0).filter(date__year=year, date__month=month).select_related("project_m", "user_m").order_by("date")
     months = HourEntry.objects.filter(user_m=person).exclude(incurred_hours=0).dates("date", "month", order="DESC")
-    return render(request, "person.html", {"month_or_week": "month", "person": person, "hour_entries": entries, "months": months, "report_number": month, "year": year, "stats": calculate_entry_stats(entries, [])})
-
-
-@login_required
-def person_details_week(request, year, week, user_guid):
-    year = int(year)
-    week = int(week)
-    week_start = date_utils.week_start_date(year, week)
-    week_end = date_utils.week_end_date(year, week)
-    person = get_object_or_404(FeetUser, guid=user_guid)
-    entries = person.hourentry_set.exclude(incurred_hours=0).filter(date__year=year, date__range=(week_start, week_end)).select_related("project_m", "user_m").order_by("date")
-    months = HourEntry.objects.filter(user_m=person).exclude(incurred_hours=0).dates("date","month", order="DESC")
-    weeks = set()
-    current_week = datetime.date.today().isocalendar()[1]
-    for month in months:
-        cal = calendar.Calendar()
-        monthweeks = cal.monthdatescalendar(month.year, month.month)
-        for weekdays in monthweeks:
-            week_number = weekdays[0].isocalendar()[1]  # Weekdays contain all seven days of the week, choose the first (Monday)
-            if week_number <= current_week:  # Do not add week numbers that are in the future
-                weeks.add((week_number, month.year))
-    weeks_sorted = sorted(weeks, key=lambda tup: (tup[1], tup[0]), reverse=True)  # Sort first by year and then by week
-    return render(request, "person.html", {"month_or_week": "week", "person": person, "hour_entries": entries, "weeks": weeks_sorted, "report_number": week, "year": year, "stats": calculate_entry_stats(entries, [])})
+    return render(request, "person.html", {"person": person, "hour_entries": entries, "months": months, "month": month, "year": year, "stats": calculate_entry_stats(entries, [])})
 
 
 @login_required
@@ -272,10 +244,8 @@ def people_list(request):
             person["bill_rate_avg_billable"] = person["billable"]["incurred_money"] / person["billable"]["incurred_hours"]
     return render(request, "people.html", {"people": people_data, "year": year, "month": month})
 
-
 def parse_date(date_string):
     return datetime.datetime.strptime(date_string, "%Y-%m-%d").date()
-
 
 @login_required
 def queue_update(request):
@@ -306,18 +276,15 @@ def queue_update(request):
 
 
 @login_required
-def get_pdf(request, invoice_id=None, weekly_report_id=None):
-    if invoice_id:
+def get_pdf(request, invoice_id, pdf_type):
+    if pdf_type == "hours":
         pdf, title = generate_hours_pdf_for_invoice(request, invoice_id)
-    elif weekly_report_id:
-        pdf, title = generate_hours_pdf_for_weekly_report(request, weekly_report_id)
     else:
         return HttpResponseBadRequest("Invalid PDF type")
 
     response = HttpResponse(pdf, content_type="application/pdf")
     response['Content-Disposition'] = u'attachment; filename="Hours for %s.pdf"' % title
     return response
-
 
 @login_required
 def frontpage(request):
@@ -354,17 +321,6 @@ def invoice_hours(request, invoice_id):
 
 
 @login_required
-def weekly_report_hours(request, weekly_report_id):
-    weekly_report = get_object_or_404(WeeklyReport, weekly_report_id=weekly_report_id)
-    entries = HourEntry.objects.filter(weekly_report=weekly_report).filter(incurred_hours__gt=0).select_related("user_m")
-    context = {
-        "weekly_report": weekly_report,
-        "entries": entries,
-    }
-    return render(request, "weekly_report_hours.html", context)
-
-
-@login_required
 def projects_list(request):
     projects = Project.objects.annotate(incurred_money=Sum("invoice__incurred_money"), incurred_hours=Sum("invoice__incurred_hours")).exclude(incurred_hours=0).prefetch_related("admin_users")
     show_only = request.GET.get("show_only", "").split(",")
@@ -392,14 +348,6 @@ def project_details(request, project_id):
     invoices = Invoice.objects.filter(project_m=project).exclude(Q(incurred_hours=0) & Q(incurred_money=0))
     filters = ProjectsFilter(request.GET, queryset=invoices)
     table = ProjectDetailsTable(filters.qs)
-
-    if request.method == "POST":
-        enable_notifications = request.POST.get("weeklyReportEnableNotifications", False) in (True, "true", "on")
-        project.enable_weekly_notifications = enable_notifications
-        project.save()
-        messages.add_message(request, messages.INFO, 'Slack notifications enabled.' if enable_notifications else 'Slack notifications disabled.')
-        return HttpResponseRedirect(reverse("project", args=[project_id]))
-
     RequestConfig(request, paginate={
         'per_page': 250
     }).configure(table)
@@ -441,7 +389,6 @@ def hours_charts(request):
     hours_per_month_data = [["Date", "Incurred hours"]] + [["%s-%s" % (entry["month"].year, entry["month"].month), entry["hours"]] for entry in entries]
     linecharts.append(("incurred_hours", "Incurred hours per month", json.dumps(hours_per_month_data)))
     return render(request, "hours_charts.html", {"treemap_charts": treemaps, "line_charts": linecharts, "calendar_charts": calendar_charts})
-
 
 @login_required
 def people_charts(request):
@@ -499,6 +446,7 @@ def project_charts(request, project_id):
     calendar_charts.append(("hours_calendar", "Incurred hours per day", "Hours", hours_calendar_data))
     calendar_charts.append(("money_calendar", "Incurred billing per day", "Money", money_calendar_data))
 
+
     entries = HourEntry.objects.filter(project_m=project).filter(calculated_is_billable=True).annotate(month=TruncMonth("date")).order_by("month").values("month").annotate(hours=Sum("incurred_hours")).annotate(money=Sum("incurred_money")).values("month", "hours", "money")
     monthly_avg_billing = [["Date", "Bill rate avg"]] + [["%s-%s" % (entry["month"].year, entry["month"].month), entry["money"] / entry["hours"]] for entry in entries]
     linecharts.append(("billing_rate_avg", "Billing rate avg", json.dumps(monthly_avg_billing)))
@@ -520,7 +468,6 @@ def invoice_charts(request, invoice_id):
         previous_invoices = []
     first_day = datetime.datetime.strptime(request.GET.get("first_day", invoice.month_start_date.strftime("%Y-%m-%d")), "%Y-%m-%d")
     last_day = datetime.datetime.strptime(request.GET.get("last_day", invoice.month_end_date.strftime("%Y-%m-%d")), "%Y-%m-%d")
-
     def get_chart_data(queryset):
         d = defaultdict(float)
         for item in queryset:
@@ -638,13 +585,10 @@ def invoice_page(request, invoice_id, **_):
     except Comments.DoesNotExist:
         latest_comments = None
 
+
     previous_invoices = []
-    recent_weekly_report = None
     if invoice.project_m:
         previous_invoices = Invoice.objects.filter(project_m=invoice.project_m)
-        weekly_reports = WeeklyReport.objects.filter(project_m=invoice.project_m)
-        if weekly_reports:
-            recent_weekly_report = weekly_reports[0]
 
     context = {
         "today": today,
@@ -654,9 +598,7 @@ def invoice_page(request, invoice_id, **_):
         "invoice": invoice,
         "previous_invoices": previous_invoices,
         "recent_invoice": abs((datetime.date.today() - datetime.date(invoice.year, invoice.month, 1)).days) < 60,
-        "recent_weekly_report": recent_weekly_report
     }
-
     context.update(entry_data)
 
     previous_invoice_month = invoice.month - 1
@@ -667,70 +609,8 @@ def invoice_page(request, invoice_id, **_):
     try:
         last_month_invoice = Invoice.objects.get(project=invoice.project, client=invoice.client, year=previous_invoice_year, month=previous_invoice_month)
         context["last_month_invoice"] = last_month_invoice
-        context["diff_last_month"] = compare_invoices(last_month_invoice, invoice)
+        context["diff_last_month"] = last_month_invoice.compare(invoice)
     except Invoice.DoesNotExist:
         pass
 
     return render(request, "invoice_page.html", context)
-
-
-@login_required
-def weekly_report_page(request, weekly_report_id, **_):
-    weekly_report = get_object_or_404(WeeklyReport, weekly_report_id=weekly_report_id)
-
-    today = datetime.date.today()
-
-    entries = HourEntry.objects.filter(weekly_report=weekly_report).filter(incurred_hours__gt=0)
-    entry_data = calculate_weekly_entry_stats(entries)
-
-    if request.method == "POST":
-        comment = WeeklyReportComments(checked=request.POST.get("weeklyReportChecked", False) in (True, "true", "on"),
-                                       user=request.user.email,
-                                       weekly_report=weekly_report)
-        comment.save()
-        weekly_report.is_approved = comment.checked
-        messages.add_message(request, messages.INFO, 'Saved.')
-        weekly_report.save()
-        weekly_report.update_state(comment=comment)
-        weekly_report.save()
-        return HttpResponseRedirect(reverse("weekly_report", args=[weekly_report.weekly_report_id]))
-
-    try:
-        latest_comments = WeeklyReportComments.objects.filter(weekly_report=weekly_report).latest()
-    except WeeklyReportComments.DoesNotExist:
-        latest_comments = None
-
-    previous_weekly_reports = []
-
-    if weekly_report.project_m:
-        previous_weekly_reports = WeeklyReport.objects.filter(project_m=weekly_report.project_m)
-
-    context = {
-        "today": today,
-        "entries": entries,
-        "weekly_report": weekly_report,
-        "form_data": latest_comments,
-        "previous_weekly_reports": previous_weekly_reports,
-    }
-
-    context.update(entry_data)
-
-    previous_weekly_report_week = weekly_report.week - 1
-    previous_weekly_report_year = weekly_report.year
-    if previous_weekly_report_week == 0:
-        if date_utils.weeks_in_a_year(previous_weekly_report_year-1) == 53:
-            previous_weekly_report_week = 53
-        else:
-            previous_weekly_report_week = 52
-        previous_weekly_report_year -= 1
-    try:
-        last_weeks_report = WeeklyReport.objects.get(
-            project_m=weekly_report.project_m,
-            year=previous_weekly_report_year,
-            week=previous_weekly_report_week)
-        context["last_weeks_report"] = last_weeks_report
-        context["diff_last_week"] = compare_invoices(last_weeks_report, weekly_report)
-    except WeeklyReport.DoesNotExist:
-        pass
-
-    return render(request, "weekly_report_page.html", context)
